set C# version to 8.0 in explicit way so we it can be build using netcoreapp3.1 SDK and does not get accidentaly broken over time

+++ src/benchmarks/micro/libraries/Common/AlignedMemory.cs
@@ -23,7 +23,7 @@ namespace System
             _length = length;
         }
 
-        public static unsafe AlignedMemory Allocate(nuint length, nuint alignment)
+        public static unsafe AlignedMemory Allocate(uint length, uint alignment)
             => new AlignedMemory(NativeMemory.AlignedAlloc(length, alignment), (int)length);
 
         public bool IsDisposed => _disposed;
